+++ dace/libraries/blas/blas_helpers.py
@@ -23,13 +23,13 @@ def to_blastype(dtype):
 
 
 def get_gemm_opts(a_strides, b_strides, c_strides) -> Dict[str, Any]:
-    """ 
+    """
     Returns GEMM argument order, transposition, and leading dimensions
-    based on column-major storage from dace arrays. 
-    :param a: Data descriptor for the first matrix. 
+    based on column-major storage from dace arrays.
+    :param a: Data descriptor for the first matrix.
     :param b: Data descriptor for the second matrix.
     :param c: Data descriptor for the output matrix.
-    :return: A dictionary with the following keys: swap (if True, a and b 
+    :return: A dictionary with the following keys: swap (if True, a and b
              should be swapped); lda, ldb, ldc (leading dimensions); ta, tb
              (whether GEMM should be called with OP_N or OP_T).
     """
@@ -141,3 +141,28 @@ def get_gemm_opts(a_strides, b_strides, c_strides) -> Dict[str, Any]:
         raise Exception("sCM or sCN should be 1")
 
     return opts[optA + optB + optC]
+
+
+def _get_matmul_inputs(node, state, sdfg):
+    """Returns the matrix multiplication input edges, arrays, and shape."""
+    res_a = None
+    res_b = None
+    for edge in state.in_edges(node):
+        if edge.dst_conn in ["_a", "_b"]:
+            subset = dc(edge.data.subset)
+            squeezed = subset.squeeze()
+            size = subset.size()
+            outer_array = sdfg.data(
+                dace.sdfg.find_input_arraynode(state, edge).data)
+            strides = [
+                s for i, s in enumerate(outer_array.strides) if i in squeezed
+            ]
+            res = edge, outer_array, size, strides
+            if edge.dst_conn == "_a":
+                res_a = res
+            else:
+                res_b = res
+    if res_a is None or res_b is None:
+        raise ValueError("Matrix multiplication input connectors \"_a\" and "
+                         "\"_b\" not found.")
+    return res_a, res_b
