+++ dace/libraries/blas/nodes/__init__.py
@@ -1,4 +1,5 @@
 from .dot import Dot
 from .matmul import MatMul
 from .gemm import Gemm
+from .batched_matmul import BatchedMatMul
 from .transpose import Transpose
